Create a branch in xic-build to check for sys gradle
Some systems have gradle installed by the package manager, and we can
just call gradle <task> instead of ./gradlew <task>. Made this change
in the xic-build script to care for in the VM.

(apt-installing gradle on my local machine results in numerous errors,
so I just let IDEA maintain a local copy). <-- Might need to add the
local copy to the git repo, and the zip we submit.