Fix issue: imm64 not allowed anywhere except as operands to mov
The toASM function for an IRConst c now checks if the c is less than 32
bits. If it is, then c is wrapped inside an ExprConst. If not, then
c is moved to a temp t, and t is returned from the function. This
fixes the issue where the binops/cmp would have directly used imm64,
which is not allowed in x86_64