Replace forceful callee-save reg pushing with scan-based adaptive method
The new arrangement only pushes callee-save regs when it needs to. That
is, if no callee-save regs are used in a function body, no need to spill
them to the stack.

Also, using ENTER, LEAVE instead of explicit push rbp etc. Is cleaner,
and works better with sub rsp reduction optimization.